+++ nebula-simulator/src/main/java/edu/illinois/cs/srg/sim/cluster/EndEvent.java
@@ -13,8 +13,8 @@ public class EndEvent {
     return Long.parseLong(event[1]);
   }
 
-  public static long getIndex(String[] event) {
-    return Long.parseLong(event[2]);
+  public static int getIndex(String[] event) {
+    return Integer.parseInt(event[2]);
   }
 
   public static long getStartTime(String[] event) {
